Reworked transport layer. - Separate Classes for the version negotiation (according to strategy pattern) - Separate Classes for authentication supporting SASL (according to strategy pattern) - Different minor bugs found and fixed. - Error handling at transport level improved (minor errors are now allowed in transport phase and not fatal) - SASL PLAIN and EXTERNAL Client/Server implemented according to Java SASL API (Factory and Provider still missing, so only for experimental use)

+++ core/message/src/main/java/org/ietf/nea/pt/value/enums/PtTlsSaslResultEnum.java
@@ -91,7 +91,7 @@ public enum PtTlsSaslResultEnum {
      * @param id the result ID
      * @return a result or null
      */
-    public static PtTlsSaslResultEnum fromId(final long id) {
+    public static PtTlsSaslResultEnum fromId(final int id) {
 
         if (id == SUCCESS.id) {
             return SUCCESS;
